fix exercise video byhhusang

+++ Androi_Client/app/src/main/java/com/example/app_selfcare/Adapter/WorkoutSliderAdapter.java
@@ -41,7 +41,7 @@ public class WorkoutSliderAdapter extends RecyclerView.Adapter<WorkoutSliderAdap
         ExerciseResponse exercise = exercises.get(position);
         
         holder.tvWorkoutName.setText(exercise.getExerciseName());
-        holder.tvCalories.setText((int) exercise.getCaloriesPerMinute() + " cal/phút");
+        holder.tvCalories.setText((double) exercise.getCaloriesPerMinute() + " cal/phút");
         holder.tvDifficulty.setText(mapDifficulty(exercise.getDifficultyLevel()));
         
         // Load image
